Functional Donation

+++ HelpPawsGuevara/src/main/java/co/edu/cue/HelpPawsGuevara/mapping/dtos/FoundationDto.java
@@ -1,6 +1,6 @@
 package co.edu.cue.HelpPawsGuevara.mapping.dtos;
 
-public record FoundationDto ( Long Id,
+public record FoundationDto ( int Id,
          String name,
          String address,
          String phone) {
