+++ HelpPawsGuevara/src/main/java/co/edu/cue/HelpPawsGuevara/mapping/dtos/UserDto.java
@@ -1,6 +1,6 @@
 package co.edu.cue.HelpPawsGuevara.mapping.dtos;
 
-public record UserDto ( Long id,
+public record UserDto ( int id,
          String name,
          String email,
          String password){
